+++ dorm_postgres_database/example/lib/main.dart
@@ -52,7 +52,7 @@ Future<void> main() async {
     );
 
     final List<User> users = await dorm.users.repository.peekAll(
-      Filter.text('Ad', key: UserEntity.fields.name.fieldName),
+      Filter.text('Ad', field: UserEntity.fields.name),
     );
     final List<Join<User, Post>> posts = await dorm.relations.users.posts
         .peekAll();